+++ src/Microservice/Microservice.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Mtarld\ApiPlatformMsBundle\Microservice;
+
+use Symfony\Component\Validator\Constraints as Assert;
+
+class Microservice
+{
+    /**
+     * @Assert\Url
+     */
+    private $baseUri;
+
+    public function __construct(string $baseUri)
+    {
+        $this->baseUri = $baseUri;
+    }
+
+    public function getBaseUri(): string
+    {
+        return $this->baseUri;
+    }
+}